fastapi-app-insights/pyproject.toml
[tool.poetry]
name = "fastapi-app-insights"
version = "0.1.0"
description = "Fast API demo with app insights"
authors = ["Anthony Shaw"]
packages = [
    { include = "example_app", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "*"
uvicorn = "^0.14.0"
opentelemetry-api = "^1.4.1"
opentelemetry-sdk = "^1.4.1"
opentelemetry-instrumentation-fastapi = "^0.24b0"
azure-monitor-opentelemetry-exporter = {version = "^1.0.0-beta.4", allow-prereleases = true}
rich = "^10.9.0"
mimesis = "^4.1.3"
tortoise-orm = "^0.17.7"
opentelemetry-exporter-richconsole = {path = "../../opentelemetry-python-contrib/exporter/opentelemetry-exporter-richconsole"}
opentelemetry-instrumentation-tortoiseorm = {path = "../../opentelemetry-python-contrib/instrumentation/opentelemetry-instrumentation-tortoiseorm"}
opentelemetry-instrumentation-jinja2 = "^0.24b0"
Jinja2 = "2.7"
aiofiles = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.4"
pytest-asyncio = "*"
black = "*"